excel-to-word-report-generator/excel-to-word-report-generator(direct-reading)/excel-to-word-report-generator(direct-reading)/requirements.txt
python-docx==1.1.2
openpyxl==3.1.2
lxml==5.2.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

